auth logic

+++ routes/web.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\Dashboard;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\Authentication;
 use App\Http\Controllers\UserController;
 use App\Http\Controllers\LogbookController;
@@ -19,17 +20,27 @@ use App\Http\Controllers\DepartmentController;
 |
 */
 
-Route::get('/',[Dashboard::class, 'index'])->name('dashboard');
-Route::get('/dashboard',[Dashboard::class, 'index']);
+Route::get('/login',[AuthController::class, 'login'])->name('login')->middleware('guest');
+Route::post('/login',[AuthController::class, 'validation'])->name('validation')->middleware('guest');
+Route::post('/logout',[AuthController::class, 'logout'])->name('logout')->middleware('guest');
+
+Route::get('/register',[AuthController::class, 'register'])->name('get_register')->middleware('guest');
+Route::post('/register',[AuthController::class, 'store_register'])->name('get_register')->middleware('guest');
+
+
+Route::get('/',[Dashboard::class, 'index'])->name('dashboard')->middleware('auth');
+Route::get('/dashboard',[Dashboard::class, 'index'])->name('dashboard')->middleware('auth');
+
+//Route masih satu persatu agar tidak lupa, next development bisa di rapikan dan grouping
 
 // Users Route
-Route::get('/users',[UserController::class, 'index'])->name('get_users');
-Route::get('/add_users',[UserController::class, 'get_users_add'])->name('get_users_add');
-Route::post('/add_users',[UserController::class, 'store_user'])->name('user_add');
-Route::get('/update_users/{id}',[UserController::class, 'get_update_user'])->name('get_user_update');
-Route::put('/update_users',[UserController::class, 'update_user'])->name('user_update');
-Route::get('/delete_users',[UserController::class, 'get_delete_user'])->name('get_user_delete');
-Route::delete('/delete_users/{id}',[UserController::class, 'delete_user'])->name('user_delete');
+Route::get('/users',[UserController::class, 'index'])->name('get_users')->middleware('auth');
+Route::get('/add_users',[UserController::class, 'get_users_add'])->name('get_users_add')->middleware('auth');
+Route::post('/add_users',[UserController::class, 'store_user'])->name('user_add')->middleware('auth');
+Route::get('/update_users/{id}',[UserController::class, 'get_update_user'])->name('get_user_update')->middleware('auth');
+Route::put('/update_users',[UserController::class, 'update_user'])->name('user_update')->middleware('auth');
+Route::get('/delete_users',[UserController::class, 'get_delete_user'])->name('get_user_delete')->middleware('auth');
+Route::delete('/delete_users/{id}',[UserController::class, 'delete_user'])->name('user_delete')->middleware('auth');
 
 // Positions Route
 Route::get('/positions',[Dashboard::class, 'get_position'])->name('get_positions');
@@ -54,6 +65,8 @@ Route::delete('/delete_departments/{id}',[DepartmentController::class, 'destroy'
 
 //Logbook
 Route::get('/logbook',[LogbookController::class, 'index'])->name('get_logbook');
+Route::get('/logbook-acc/{id}',[LogbookController::class, 'get_acc'])->name('get_acc');
+Route::put('/logbook-acc',[LogbookController::class, 'acc'])->name('logbook_acc');
 Route::get('/add-logbook',[LogbookController::class, 'create'])->name('get_logbook_add');
 Route::post('/add-logbook',[LogbookController::class, 'store'])->name('logbook_add');
 Route::get('/update-logbook/{id}',[LogbookController::class, 'edit'])->name('get_logbook_update');
@@ -61,6 +74,5 @@ Route::put('/update-logbook',[LogbookController::class, 'update'])->name('logboo
 Route::get('/delete-logbook',[LogbookController::class, 'get_destroy'])->name('get_logbook_update');
 Route::delete('/delete-logbook/{id}',[LogbookController::class, 'destroy'])->name('logbook_delete');
 
-Route::get('/login',[Authentication::class, 'get_login']);
-Route::get('/register',[Authentication::class, 'get_register']);
+
